# Conflicts:
#	README.md
#	console_test.go
#	diode/diode.go
#	diode/diode_example_test.go
#	go.mod
#	go.sum
#	hlog/hlog.go
#	hlog/hlog_example_test.go
#	hlog/hlog_test.go
#	log.go
#	pkgerrors/stacktrace_test.go

+++ internal/cbor/types_test.go
@@ -284,7 +284,7 @@ func BenchmarkAppendInt(b *testing.B) {
 				case 23:
 					_ = enc.AppendInt32(buf, int32(str.val))
 				case 24:
-					_ = enc.AppendInt64(buf, int64(str.val))
+					_ = enc.AppendInt64(buf, str.val)
 				}
 			}
 		})
